Create Chatbot
In order to successfully run this code on any PC;
1. Make sure to install Python.
Ensure that your IDE is configured to use the correct Python environment where all the required packages are installed.
2. Install NLTK:
The nltk library is very important to run the program. You can install it via pip if it’s not already installed:
Use Command: 'pip install nltk' on command prompt
3. Download NLTK Data:
This script includes a line to download the necessary NLTK data:
Code: nltk.download('punkt')
This will download the 'punkt' tokenizer models required to process text.
4. Run the Script:
You can run the script directly in an IDE that is compatible with python. The script is interactive, meaning it will prompt you for input and it will respond based on the conversation pairs defined.
Make sure your Spyder console is ready to accept input, as the input() function will expect user interaction.
5. Have an internet Connection:
You will need an active internet connection when you first run the script to download the NLTK data.
6. Consideration this to Run:
Since the script uses a while loop to handle user input until the user types "quit", ensure you're prepared to interact with the chatbot through the console.